@@ -1,113 +1,84 @@
 {
-    "name": "acip-web",
-    "version": "0.0.0",
-    "description": "Description for acip web",
+    "name": "ng-web",
+    "version": "1.0.0",
+    "description": "Description for ng web",
     "private": true,
     "cacheDirectories": [
         "node_modules"
     ],
     "dependencies": {
-        "@angular/common": "^4.1.1",
-        "@angular/compiler": "^4.1.1",
-        "@angular/core": "^4.1.1",
-        "@angular/forms": "^4.1.1",
-        "@angular/http": "^4.1.1",
-        "@angular/platform-browser": "^4.1.1",
-        "@angular/platform-browser-dynamic": "^4.1.1",
-        "@angular/router": "^4.1.1",
-        "@ng-bootstrap/ng-bootstrap": "^1.0.0-alpha.25",
-        "bootstrap": "^4.0.0-alpha.5",
-        "brace": "^0.10.0",
-        "core-js": "^2.4.1",
-        "crypto-js": "^3.1.9-1",
-        "echarts": "^3.5.1",
-        "jquery": "^3.1.1",
-        "jquery-mousewheel": "^3.1.13",
-        "ng2-webstorage": "^1.5.0",
-        "perfect-scrollbar": "^0.6.16",
-        "reflect-metadata": "^0.1.9",
-        "rxjs": "^5.0.3",
-        "sockjs-client": "^1.1.1",
-        "sweetalert2": "^6.4.2",
-        "tether": "^1.4.0",
-        "tether-tooltip": "^1.2.0",
-        "toastr": "^2.1.2",
-        "webstomp-client": "^1.0.3",
-        "x2js": "^3.1.0",
-        "zone.js": "^0.7.6"
+        "@angular/common": "~7.2.0",
+        "@angular/compiler": "~7.2.0",
+        "@angular/core": "~7.2.0",
+        "@angular/forms": "~7.2.0",
+        "@angular/platform-browser": "~7.2.0",
+        "@angular/platform-browser-dynamic": "~7.2.0",
+        "@angular/router": "~7.2.0",
+        "@ng-bootstrap/ng-bootstrap": "~4.0.1",
+        "ace-diff": "~2.3.0",
+        "animate.css": "~3.7.0",
+        "bootstrap": "~4.2.1",
+        "brace": "~0.11.1",
+        "core-js": "~2.6.0",
+        "crypto-js": "~3.1.9-1",
+        "echarts": "~4.1.0",
+        "intersection-observer": "^0.5.1",
+        "jquery": "~3.3.1",
+        "jquery-mousewheel": "~3.1.13",
+        "jsencrypt": "~3.0.0-rc.1",
+        "ng-lazyload-image": "^5.0.0",
+        "ngx-webstorage": "~3.0.0-beta.14",
+        "perfect-scrollbar": "~1.4.0",
+        "popper.js": "~1.14.6",
+        "rxjs": "~6.3.3",
+        "sweetalert2": "~7.33.1",
+        "toastr": "~2.1.4",
+        "zone.js": "~0.8.26"
     },
     "devDependencies": {
-        "@types/crypto-js": "^3.1.33",
-        "@types/echarts": "^0.0.5",
-        "@types/jasmine": "^2.5.41",
-        "@types/jquery": "^2.0.40",
-        "@types/jquery-mousewheel": "^3.1.5",
-        "@types/jquery.slimscroll": "^1.3.30",
-        "@types/node": "^7.0.5",
-        "@types/selenium-webdriver": "^2.53.39",
-        "@types/tether": "^1.4.0",
-        "@types/toastr": "^2.1.32",
-        "add-asset-html-webpack-plugin": "^2.0.1",
-        "angular2-template-loader": "0.6.0",
-        "autoprefixer": "^6.7.3",
-        "awesome-typescript-loader": "^3.0.0-beta.18",
-        "bootstrap-loader": "^2.0.0",
-        "browser-sync": "^2.18.6",
-        "browser-sync-webpack-plugin": "^1.1.3",
-        "codelyzer": "^3.0.1",
-        "copy-webpack-plugin": "^4.0.1",
-        "css-loader": "^0.26.1",
-        "del": "^2.2.2",
-        "event-stream": "^3.3.4",
-        "exports-loader": "^0.6.3",
-        "extract-text-webpack-plugin": "^2.1.0",
-        "file-loader": "^0.10.0",
-        "html-loader": "^0.4.4",
-        "html-webpack-plugin": "^2.26.0",
-        "image-webpack-loader": "^3.1.0",
-        "jasmine-core": "^2.5.2",
-        "jasmine-reporters": "^2.2.0",
-        "karma": "^1.4.0",
-        "karma-chrome-launcher": "^2.0.0",
-        "karma-coverage": "^1.1.1",
-        "karma-jasmine": "^1.1.0",
-        "karma-junit-reporter": "^1.2.0",
-        "karma-phantomjs-launcher": "^1.0.2",
-        "karma-remap-istanbul": "^0.6.0",
-        "karma-sourcemap-loader": "^0.3.7",
-        "karma-webpack": "^2.0.1",
-        "lazypipe": "^1.0.1",
-        "lodash": "^4.17.4",
-        "map-stream": "^0.0.6",
-        "ng-router-loader": "^2.1.0",
-        "node-sass": "^4.5.0",
-        "optimize-js-plugin": "^0.0.4",
-        "phantomjs-prebuilt": "^2.1.14",
-        "postcss-loader": "^1.2.0",
-        "protractor": "^5.1.1",
-        "protractor-jasmine2-screenshot-reporter": "^0.3.2",
-        "proxy-middleware": "^0.15.0",
-        "raw-loader": "^0.5.1",
-        "resolve-url-loader": "^2.0.0",
-        "run-sequence": "^1.2.2",
-        "sass-loader": "^6.0.3",
-        "sass-resources-loader": "^1.2.0",
-        "source-map-loader": "^0.1.6",
-        "sourcemap-istanbul-instrumenter-loader": "^0.2.0",
-        "string-replace-webpack-plugin": "^0.1.3",
-        "style-loader": "^0.13.1",
-        "to-string-loader": "^1.1.5",
-        "ts-node": "^3.0.2",
-        "tslint": "^4.3.1",
-        "tslint-loader": "^3.3.0",
-        "typescript": "^2.2.1",
-        "url-loader": "^0.5.8",
-        "webpack": "^2.2.0",
-        "webpack-dev-server": "^2.2.0",
-        "webpack-merge": "^4.1.0",
-        "webpack-visualizer-plugin": "^0.1.10",
-        "write-file-webpack-plugin": "^4.0.0",
-        "yargs": "^6.6.0"
+        "@angular-devkit/build-optimizer": "~0.12.0",
+        "@angular/compiler-cli": "~7.2.0",
+        "@angular/language-service": "~7.2.0",
+        "@ngtools/webpack": "~7.2.0",
+        "@types/crypto-js": "~3.1.43",
+        "@types/echarts": "~4.1.1",
+        "@types/jquery": "~3.3.22",
+        "@types/jquery-mousewheel": "~3.1.8",
+        "@types/node": "^10.12.0",
+        "@types/tether": "~1.4.4",
+        "@types/toastr": "~2.1.35",
+        "autoprefixer": "~9.4.4",
+        "browser-sync": "~2.26.0",
+        "browser-sync-webpack-plugin": "~2.2.2",
+        "codelyzer": "~4.5.0",
+        "copy-webpack-plugin": "~4.6.0",
+        "css-loader": "~2.1.0",
+        "del": "~3.0.0",
+        "exports-loader": "~0.7.0",
+        "file-loader": "~3.0.1",
+        "html-loader": "~0.5.5",
+        "html-webpack-plugin": "~3.2.0",
+        "imports-loader": "~0.8.0",
+        "inline-manifest-webpack-plugin": "~4.0.1",
+        "mini-css-extract-plugin": "~0.5.0",
+        "node-sass": "~4.11.0",
+        "postcss-loader": "~3.0.0",
+        "raw-loader": "~1.0.0",
+        "sass-loader": "~7.1.0",
+        "sass-resources-loader": "~2.0.0",
+        "source-map-loader": "~0.2.1",
+        "style-loader": "~0.23.1",
+        "to-string-loader": "~1.1.5",
+        "ts-node": "~7.0.1",
+        "tslint": "~5.12.0",
+        "tslint-loader": "~3.6.0",
+        "typescript": "~3.2.2",
+        "webpack": "~4.28.0",
+        "webpack-bundle-analyzer": "^3.0.3",
+        "webpack-cli": "~3.2.1",
+        "webpack-dev-server": "~3.1.14",
+        "webpack-merge": "~4.2.1",
+        "yargs": "~12.0.2"
     },
     "engines": {
         "node": ">=6.9.0"
