@@ -1,9 +1,10 @@
 import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, CanActivateChild, Router, RouterStateSnapshot } from '@angular/router';
-import { LocalStorageService } from 'ng2-webstorage';
-import { Observable } from 'rxjs/Observable';
+import { LocalStorageService } from 'ngx-webstorage';
+import { Observable } from 'rxjs';
+import { isPresent } from 'share';
 
-@Injectable()
+@Injectable({ providedIn: 'root' })
 export class AuthGuardService implements CanActivate, CanActivateChild {
 
     constructor(private router: Router, private storage: LocalStorageService) {
@@ -13,7 +14,7 @@ export class AuthGuardService implements CanActivate, CanActivateChild {
         | Promise<boolean>
         | boolean {
 
-        if (this.storage.retrieve('token')) {
+        if (isPresent(this.storage.retrieve('token')) && isPresent(this.storage.retrieve('user'))) {
             return true;
         }
         this.router.navigate(['login']);
