@@ -2,10 +2,10 @@ const del = require('del');
 const path = require('path');
 const argv = require('yargs').argv;
 
-if(argv.p) {
-    del.sync([path.resolve('dist/**')]);
-    del.sync([path.resolve('dll/**')]);
-} else {
-    del.sync([path.resolve('.tmp/**')]);
-}
+del.sync([path.resolve('.tmp/**')]);
+del.sync([path.resolve('.awcache/**')]);
 
+if (argv.p) {
+    del.sync([path.resolve('dist/**')]);
+    del.sync([path.resolve('aot/**')]);
+}
