@@ -1,103 +1,63 @@
-const webpack = require('webpack');
-const CommonsChunkPlugin = require('webpack/lib/optimize/CommonsChunkPlugin');
-const CopyWebpackPlugin = require('copy-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-const StringReplacePlugin = require('string-replace-webpack-plugin');
-const AddAssetHtmlPlugin = require('add-asset-html-webpack-plugin');
-const TsConfigPathsPlugin = require('awesome-typescript-loader').TsConfigPathsPlugin;
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const DefinePlugin = require('webpack/lib/DefinePlugin');
-const path = require('path');
+const ProvidePlugin = require('webpack/lib/ProvidePlugin');
+const InlineManifestWebpackPlugin = require('inline-manifest-webpack-plugin');
+const CopyWebpackPlugin = require('copy-webpack-plugin');
+const AngularCompilerPlugin = require('@ngtools/webpack').AngularCompilerPlugin;
+const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
 
-const config = require('./config');
+const helpers = require('./helpers');
 
-module.exports = function (options) {
+module.exports = function(options) {
+    const isProd = options.env === 'prod';
+    const metadata = Object.assign({}, {
+        HMR: true,
+        AOT: false,
+        WATCH: true,
+        tsConfigPath: 'tsconfig.json'
+    }, options.metadata || {});
+
+    const jsLoaders = {
+        test: /(?:\.ngfactory\.js|\.ngstyle\.js|\.ts)$/,
+        use: metadata.AOT ? [{
+            loader: '@angular-devkit/build-optimizer/webpack-loader',
+            options: {
+                sourceMap: false
+            }
+        }, '@ngtools/webpack'] : ['@ngtools/webpack']
+    };
+    const entry = {
+        polyfills: helpers.root('src/polyfills.ts'),
+        global: helpers.root('src/assets/scss/global.scss'),
+        main: helpers.root('src/app.main.ts')
+    };
+
     return {
-        entry: {
-            'polyfills': './src/app/polyfills',
-            'global': './src/assets/scss/global.scss',
-            'main': './src/app/app.main'
-        },
+        entry: entry,
         resolve: {
-            extensions: ['.ts', '.js'],
-            modules: ['node_modules'],
-            plugins: [
-                new TsConfigPathsPlugin()
-            ]
+            mainFields: ['browser', 'module', 'main'],
+            extensions: ['.ts', '.js', '.json'],
+            modules: [helpers.root('node_modules')]
         },
         module: {
             rules: [
-                {
-                    test: /\.pdf/,
-                    loaders: ['url-loader?limit=10000']
-                },
-                {
-                    test: /\.ts$/,
-                    loaders: ['ng-router-loader', 'angular2-template-loader', 'awesome-typescript-loader']
-                },
+                jsLoaders,
                 {
                     test: /\.html$/,
-                    loader: ['html-loader'],
+                    loader: 'html-loader',
+                    options: {
+                        minimize: true,
+                        caseSensitive: true,
+                        removeAttributeQuotes: false,
+                        minifyJS: false,
+                        minifyCSS: false
+                    },
                     exclude: [/src\/index\.html/]
-                },
-                {
-                    test: /\.scss$/,
-                    use: [
-                        'to-string-loader',
-                        'css-loader?sourceMap',
-                        'postcss-loader?sourceMap',
-                        'sass-loader?sourceMap',
-                        {
-                            loader: 'sass-resources-loader',
-                            options: {
-                                resources: [
-                                    './src/assets/scss/variables.scss',
-                                    './config/sass-resources.scss'
-                                ]
-                            }
-                        }],
-                    exclude: /global\.scss/
-                },
-                {
-                    test: /global\.scss/,
-                    use: ExtractTextPlugin.extract({
-                        fallback: "style-loader",
-                        use: [
-                            "css-loader?sourceMap",
-                            'postcss-loader?sourceMap',
-                            'sass-loader?sourceMap',
-                            {
-                                loader: 'sass-resources-loader',
-                                options: {
-                                    resources: [
-                                        './src/assets/scss/variables.scss',
-                                        './config/sass-resources.scss'
-                                    ]
-                                }
-                            }
-                        ]
-                    })
                 },
                 {
                     test: /\.(jpe?g|png|gif)$/i,
                     loaders: [
-                        'file-loader?hash=sha512&digest=hex&name=assets/images/[hash].[ext]',
-                        {
-                            loader: 'image-webpack-loader',
-                            query: {
-                                progressive: true,
-                                pngquant: {
-                                    quality: '65-90',
-                                    speed: 4
-                                },
-                                optipng: {
-                                    optimizationLevel: 7
-                                },
-                                gifsicle: {
-                                    interlaced: false
-                                }
-                            }
-                        }
+                        'file-loader?hash=sha512&digest=hex&name=assets/images/[hash].[ext]'
                     ]
                 },
                 {
@@ -107,60 +67,83 @@ module.exports = function (options) {
                     ]
                 },
                 {
-                    test: /app.constants.ts$/,
-                    loader: StringReplacePlugin.replace({
-                        replacements: [
-                            {
-                                pattern: /\/\* @replace (\w*?) \*\//ig,
-                                replacement: function (match, p1) {
-                                    return `_${p1} = ${config[p1]};`;
-                                }
-                            }
-                        ]
-                    })
+                    test: /404\.html$/,
+                    use: [{
+                        loader: 'html-loader',
+                        options: {
+                            minimize: true
+                        }
+                    }]
                 }
             ]
         },
+        optimization: {
+            splitChunks: {
+                chunks: 'async',
+                minSize: 30000,
+                minChunks: 1,
+                maxAsyncRequests: 5,
+                maxInitialRequests: 3,
+                name: true,
+                cacheGroups: {
+                    vendors: {
+                        test: /[\\/]node_modules[\\/]/,
+                        priority: -10,
+                        chunks: 'all'
+                    },
+                    default: {
+                        minChunks: 2,
+                        priority: -20,
+                        reuseExistingChunk: true
+                    }
+                }
+            },
+            runtimeChunk: {
+                name: entry => `runtime~${entry.name}`
+            }
+        },
         plugins: [
+            new BundleAnalyzerPlugin(),
             new DefinePlugin({
                 ENV: require('../config.json')[options.env]
             }),
-            new CommonsChunkPlugin({
-                names: ['manifest', 'polyfills'].reverse()
-            }),
-            new webpack.DllReferencePlugin({
-                context: './',
-                manifest: require(path.resolve(`${config.dllPath}/vendor.json`)),
-            }),
-            new CopyWebpackPlugin([
-                {from: './src/favicon.ico', to: 'favicon.ico'},
-                {from: './src/robots.txt', to: 'robots.txt'}
-            ]),
-            new webpack.ProvidePlugin({
-                $: "jquery",
-                jQuery: "jquery",
-                Tooltip: 'tether-tooltip',
-                toastr: 'toastr'
+            new ProvidePlugin({
+                $: 'jquery',
+                jQuery: 'jquery',
+                'window.$': 'jquery',
+                'window.jQuery': 'jquery',
+                toastr: 'toastr',
+                Popper: ['popper.js', 'default'],
+                Util: 'exports-loader?Util!bootstrap/js/dist/util'
             }),
             new HtmlWebpackPlugin({
                 template: 'src/index.html',
-                chunksSortMode: 'dependency',
-                inject: true
-            }),
-            new AddAssetHtmlPlugin([
-                {
-                    filepath: path.resolve(`${config.dllPath}/vendor.dll.js`),
-                    hash: true,
-                    includeSourcemap: false
+                chunksSortMode: function(a, b) {
+                    const entryPoints = ['polyfills', 'vendor', 'main', 'global'];
+                    return entryPoints.indexOf(a.names[0]) - entryPoints.indexOf(b.names[0]);
                 },
-                {
-                    filepath: path.resolve(`${config.dllPath}/vendor.css`),
-                    typeOfAsset: 'css',
-                    hash: true,
-                    includeSourcemap: false
-                }
+                inject: 'body',
+                xhtml: true,
+                minify: isProd ? {
+                    caseSensitive: true,
+                    collapseWhitespace: true,
+                    keepClosingSlash: true
+                } : false
+            }),
+            new CopyWebpackPlugin([
+                { from: helpers.root('src/favicon.ico'), to: 'favicon.ico' },
+                { from: helpers.root('src/404.html'), to: '404.html' }
             ]),
-            new StringReplacePlugin()
+            new AngularCompilerPlugin({
+                tsConfigPath: metadata.tsConfigPath,
+                mainPath: entry.main,
+                sourceMap: !isProd,
+                skipCodeGeneration: !metadata.AOT,
+                hostReplacementPaths: {
+                    'src/environments/environment.ts': `src/environments/environment.${isProd ? 'prod.' : ''}ts`
+                }
+            }),
+            new InlineManifestWebpackPlugin()
         ]
     };
 };
