@@ -1,15 +1,6 @@
-// DO NOT EDIT THIS FILE, EDIT THE WEBPACK COMMON CONFIG INSTEAD, WHICH WILL MODIFY THIS FILE
-let _VERSION = '0.0.0'; // This value will be overwritten by webpack
-let _DEBUG_INFO_ENABLED = true; // This value will be overwritten by webpack
-/* @replace VERSION */
-/* @replace DEBUG_INFO_ENABLED */
-export const VERSION = _VERSION;
-export const DEBUG_INFO_ENABLED = _DEBUG_INFO_ENABLED;
-
-
-export class HeaderMessage {
-    static DELETE_MESSAGE = 'Delete-Message';
-    static SHOW_ERROR_MESSAGE = 'Show-Error-Message';
-    static TRANSFORM_TO_JSON = 'Transform-To-Json';
-    static BLOCK_MESSAGE = 'Block-Message';
+export class Headers {
+    static DELETE = 'Delete-Message';
+    static SHOW_ERROR = 'Show-Error-Message';
+    static NEED_AUTHORIZATION = 'NEED_AUTHORIZATION';
+    static BLOCK = 'Block-Message';
 }
