@@ -1,2 +1,1 @@
-export { AuthGuardService, ProductGuardService } from './providers';
-export { IPageService, Page, PageComponent } from './page';
+export { AuthGuardService } from './providers';
