@@ -1,3 +1,2 @@
-export { requestOptionsProvider } from './http.options'
 export { AuthGuardService } from './auth.guard.service';
-export { ProductGuardService } from './product.guard.service';
+export { IslamicCivilI18n } from './date.datepickerI18n';
