@@ -1,4 +1,3 @@
 export { AceEditorDirective } from './ace';
-export { TetherSelectDirective } from './select/select';
 export { EchartsDirective }from './echarts';
-export { AutoFocusDirective } from './focus'
+export { AutoFocusDirective } from './focus';
