@@ -1,16 +1,12 @@
 import { CommonModule } from '@angular/common';
-import { ModuleWithProviders, NgModule } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { EventManager } from 'share';
-import { AceEditorDirective, AutoFocusDirective, EchartsDirective, TetherSelectDirective } from './directives';
-import { FormModule } from './form/form.module';
+import { AceEditorDirective, AutoFocusDirective, EchartsDirective } from './directives';
 import { JsonKeyValuePipe, KeysPipe, MaxlengthPipe, OrderByPipe } from './pipes';
-import { ScrollBarModule } from './scroll-bar/scroll-bar.module';
 
 const COMPONENTS = [];
 const DIRECTIVES = [
     AceEditorDirective,
-    TetherSelectDirective,
     EchartsDirective,
     AutoFocusDirective
 ];
@@ -20,13 +16,8 @@ const PIPES = [
     MaxlengthPipe,
     JsonKeyValuePipe
 ];
-const SERVICES = [
-    EventManager
-];
 const MODULES = [
     CommonModule,
-    ScrollBarModule,
-    FormModule
 ];
 
 @NgModule({
@@ -47,12 +38,4 @@ const MODULES = [
     ]
 })
 export class ShareModule {
-    static forRoot(): ModuleWithProviders {
-        return <ModuleWithProviders> {
-            ngModule: ShareModule,
-            providers: [
-                ...SERVICES
-            ],
-        };
-    }
 }
